Factor ContributorsList into its own file
